fix(local): add missing files at area and thermal creation (#58)
* Add missing files for create area and thermal cluster in local

* Add licences for new tests files

* Add licences for new tests files
